Use mremap for Truncate on Linux
