+++ z/file.go
@@ -152,24 +152,6 @@ func (m *MmapFile) Delete() error {
 	return os.Remove(m.Fd.Name())
 }
 
-// Truncate would truncate the mmapped file to the given size. On Linux and
-// others, we could directly just truncate the underlying file, but in Windows,
-// we can't do that. So, unmap first, then truncate, then re-map.
-func (m *MmapFile) Truncate(maxSz int64) error {
-	if err := m.Sync(); err != nil {
-		return fmt.Errorf("while sync file: %s, error: %v\n", m.Fd.Name(), err)
-	}
-	if err := Munmap(m.Data); err != nil {
-		return fmt.Errorf("while munmap file: %s, error: %v\n", m.Fd.Name(), err)
-	}
-	if err := m.Fd.Truncate(maxSz); err != nil {
-		return fmt.Errorf("while truncate file: %s, error: %v\n", m.Fd.Name(), err)
-	}
-	var err error
-	m.Data, err = Mmap(m.Fd, true, maxSz) // Mmap up to max size.
-	return err
-}
-
 // Close would close the file. It would also truncate the file if maxSz >= 0.
 func (m *MmapFile) Close(maxSz int64) error {
 	// Badger can set the m.Data directly, without setting any Fd. In that case, this should be a
